renamed files and changed some package names
