+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/controllers/PendingControllerUnitTest.java
@@ -1,4 +1,4 @@
-package com.quolance.quolance_api.unit;
+package com.quolance.quolance_api.unit.controllers;
 
 import com.quolance.quolance_api.controllers.PendingController;
 import com.quolance.quolance_api.dtos.UpdatePendingUserRequestDto;
