+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/jobs/SendTempPasswordEmailUnitTest.java
@@ -1,4 +1,4 @@
-package com.quolance.quolance_api.unit;
+package com.quolance.quolance_api.unit.jobs;
 
 import com.quolance.quolance_api.configs.ApplicationProperties;
 import com.quolance.quolance_api.entities.User;
