+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/services/business_workflow/AdminWorkflowServiceUnitTest.java
@@ -27,7 +27,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class AdminWorkflowServiceTest {
+class AdminWorkflowServiceUnitTest {
 
     @Mock
     private ProjectService projectService;
