+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/services/business_workflow/ApplicationProcessWorkflowUnitTest.java
@@ -28,7 +28,7 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class ApplicationProcessWorkflowTest {
+class ApplicationProcessWorkflowUnitTest {
 
     @Mock
     private ApplicationService applicationService;
