+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/services/business_workflow/ClientWorkflowServiceUnitTest.java
@@ -34,7 +34,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class ClientWorkflowServiceTest {
+class ClientWorkflowServiceUnitTest {
 
     @Mock
     private ProjectService projectService;
