+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/services/business_workflow/FreelancerWorkflowServiceUnitTest.java
@@ -35,7 +35,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class FreelancerWorkflowServiceTest {
+class FreelancerWorkflowServiceUnitTest {
 
     @Mock
     private ProjectService projectService;
