+++ quolance-api/src/test/java/com/quolance/quolance_api/unit/services/entity_services/ProjectServiceUnitTest.java
@@ -31,7 +31,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class ProjectServiceTest {
+class ProjectServiceUnitTest {
 
     @Mock
     private ProjectRepository projectRepository;
